+++ packages/interact/src/types/index.d.ts
@@ -1,6 +1,7 @@
 /// <reference types="./vite-env-override.d.ts" />
 /// <reference types="vite/client" />
 // Ambient virtual declare module file
+/// <reference types="../node/vite/vite-search-provider-module.d.ts" />
 /// <reference types="../node/vite/contextClientProviderModule.d.ts" />
 /// <reference types="../node/vite/contextClientProviderModule.d.ts" />
 /// <reference types="../node/vite/contextServerProviderModule.d.ts" />
@@ -11,7 +12,7 @@
 /// <reference types="../node/pages/interactPageModules.d.ts" />
 /// <reference types="../node/vite/middlewareProviderModule" />
 
-import type {SearchOptions, SearchResult, SearchResponse, SearchProvider } from "./searchProvider";
+
 import type {InteractMarkdownConfig} from "../node/markdown/conf/markdownConfig";
 import type {MiddlewareHandler, Middleware} from "../node/middlewareEngine/interactMiddleware.d.ts"
 import type {Page, Frontmatter, TocNode} from "../node/pages/interactPage";
@@ -33,11 +34,7 @@ export {
     LayoutProps,
     Frontmatter,
     TocNode,
-    PageNode,
-    SearchOptions,
-    SearchResult,
-    SearchResponse,
-    SearchProvider,
+    PageNode
 }